Remove the switchUnit function from call-api and put it in dom-manipulation file
