+++ src/modulus/call-api.js
@@ -1,30 +1,4 @@
-import { createCard, removeCard, backgroundChange } from './dom-manipulation';
-
-let unitIsCelcius = true;
-
-const switchUnit = () => {
-  const temperatureM = document.getElementById('temp');
-  const windM = document.getElementById('wind');
-  const temperatureF = document.getElementById('tempF');
-  const windF = document.getElementById('windF');
-
-  const unitSwitcher = document.getElementById('unitSwitcher');
-  unitSwitcher.addEventListener('click', () => {
-    unitIsCelcius = !unitIsCelcius;
-
-    if (!unitIsCelcius) {
-      temperatureM.style.display = 'none';
-      temperatureF.style.display = 'block';
-      windM.style.display = 'none';
-      windF.style.display = 'block';
-    } else {
-      temperatureM.style.display = 'block';
-      temperatureF.style.display = 'none';
-      windM.style.display = 'block';
-      windF.style.display = 'none';
-    }
-  });
-};
+import { createCard, removeCard, backgroundChange, switchUnit } from './dom-manipulation';
 
 const fetchWeather = async (search) => {
   try {
